UPSTREAM: fs_context flavour for submounts
This is an eventual replacement for vfs_submount() uses.  Unlike the
"mount" and "remount" cases, the users of that thing are not in VFS -
they are buried in various ->d_automount() instances and rather than
converting them all at once we introduce the (thankfully small and
simple) infrastructure here and deal with the prospective users in
afs, nfs, etc. parts of the series.

Here we just introduce a new constructor (fs_context_for_submount())
along with the corresponding enum constant to be put into fc->purpose
for those.

Change-Id: I2570fdb015e00eb9313171ea8e20f8d0ed094dc4
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
(cherry picked from commit 6a25c87bd6945d9aacd3767723b34ad245179079)